src/requirements.txt
python-chess
pandas
cairosvg